requirements.txt
discord.py==2.3.2
wavelink==2.6.4
Flask==2.3.2
aiohttp==3.8.4
PyNaCl==1.5.0
